Move conversion of oid strings to new smidump API
This way, the conversion of OID strings to OID objects takes place the
first time a smidump module gets loaded dynamically, and is not reserved
solely for smidump modules that have a corresponding MibRetriever class.
